Label login remake

+++ src/main/java/map/project/musiclibrary/service/LabelService.java
@@ -3,6 +3,7 @@ package map.project.musiclibrary.service;
 import jakarta.persistence.EntityNotFoundException;
 import map.project.musiclibrary.data.model.misc.Label;
 import map.project.musiclibrary.data.model.users.ArtistUser;
+import map.project.musiclibrary.data.model.users.UserSession;
 import map.project.musiclibrary.data.repository.ArtistUserRepository;
 import map.project.musiclibrary.data.repository.LabelRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,25 +24,33 @@ public class LabelService {
         this.artistUserRepository = artistUserRepository;
     }
 
-    public Label addLabel(String name) {
-        Label label = new Label();
-        label.setName(name);
-        label.setArtists(new ArrayList<>());
-        return labelRepository.save(label);
+    public Label addLabel(UserSession userSession, String name) {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            Label label = new Label();
+            label.setName(name);
+            label.setArtists(new ArrayList<>());
+            return labelRepository.save(label);
+        }
+        throw new SecurityException("Only admin can add labels");
     }
 
-    public void deleteLabel(Long id) {
-        Optional<Label> labelOptional = labelRepository.findById(id);
+    public void deleteLabel(UserSession userSession, String idStr) throws NumberFormatException {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            Long id = Long.parseLong(idStr);
+            Optional<Label> labelOptional = labelRepository.findById(id);
 
-        if (labelOptional.isPresent()) {
-            Label label = labelOptional.get();
-            for (ArtistUser artistUser : label.getArtists()) {
-                artistUser.setLabel(null);
-                artistUserRepository.save(artistUser);
+            if (labelOptional.isPresent()) {
+                Label label = labelOptional.get();
+                for (ArtistUser artistUser : label.getArtists()) {
+                    artistUser.setLabel(null);
+                    artistUserRepository.save(artistUser);
+                }
+                label.getArtists().clear();
+                labelRepository.deleteById(id);
             }
-            label.getArtists().clear();
-            labelRepository.deleteById(id);
+            throw new EntityNotFoundException("LabelService::Label with specified id doesn't exist");
         }
+        throw new SecurityException("Only admin can delete labels");
     }
 
     public Label save(Label label) {
@@ -52,30 +61,36 @@ public class LabelService {
         return labelRepository.findByName(name).stream().findFirst().orElse(null);
     }
 
-    public List<Label> findAll() {
-        return labelRepository.findAll();
+    public List<Label> findAll(UserSession userSession) {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            return labelRepository.findAll();
+        }
+        throw new SecurityException("Only admin can list all labels");
     }
 
-    public Label addArtist(String artistIdStr, String labelIdStr) throws NumberFormatException {
-        Long artistId = Long.parseLong(artistIdStr);
-        Long labelId = Long.parseLong(labelIdStr);
+    public Label addArtist(UserSession userSession, String artistIdStr, String labelIdStr) throws NumberFormatException {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            Long artistId = Long.parseLong(artistIdStr);
+            Long labelId = Long.parseLong(labelIdStr);
 
-        // search for the artist and label with the corresponding ids
-        Optional<ArtistUser> artistUserOptional = artistUserRepository.findById(artistId);
-        Optional<Label> labelOptional = labelRepository.findById(labelId);
+            // search for the artist and label with the corresponding ids
+            Optional<ArtistUser> artistUserOptional = artistUserRepository.findById(artistId);
+            Optional<Label> labelOptional = labelRepository.findById(labelId);
 
-        if (artistUserOptional.isPresent() && labelOptional.isPresent()) {
-            // get the artist and label with the corresponding ids
-            ArtistUser artistUser = artistUserOptional.get();
-            Label label = labelOptional.get();
-            // add artist to the labels list of artists
-            label.addArtist(artistUser);
-            artistUser.setLabel(label);
-            artistUserRepository.save(artistUser);
-            // update the label in repo
-            return labelRepository.save(label);
+            if (artistUserOptional.isPresent() && labelOptional.isPresent()) {
+                // get the artist and label with the corresponding ids
+                ArtistUser artistUser = artistUserOptional.get();
+                Label label = labelOptional.get();
+                // add artist to the labels list of artists
+                label.addArtist(artistUser);
+                artistUser.setLabel(label);
+                artistUserRepository.save(artistUser);
+                // update the label in repo
+                return labelRepository.save(label);
+            }
+
+            throw new EntityNotFoundException("LabelService::Artist or label with specified id doesn't exist");
         }
-
-        throw new EntityNotFoundException("LabelService::Artist or label with specified id doesn't exist");
+        throw new SecurityException("Only admin can add artists to labels");
     }
 }
